add test implementation of "closed" finmaps.
They work essentially by using a certificate that structurally encodes
the position of the element we are looking for and thus don't need to
return an `option t` since the certificate proves that the element is
contained in the map.
I'm not sure how practical they are to use in my code since the closed
finmaps are dependently typed.