build infrastructure to generate lemmas and generate one of the congruence lemmas using it.
Inside a lemma:
we use a dbmap (deBruijn map) which maps identifiers to the correct
deBruijn index. When generating a lemma we must manually update
it (we don't have too many nested lambda, mostly just the arguments to
the lemma at the beginning)
This allows us to easily reference the local variables and the names of
currently defined mutually inductive types/fixpoints.

Sequencing generation of multiple lemmas:
if a lemma A depends on some inductive type B we first need to generate
B and register it using MetaCoq so that A typechecks. Since I want my
generation functions to live in my own RWSE monad and not in the
TemplateMonad we keep track of the relevant definitions in an
environment that is filled while we are briefly in the
TemplateMonad (using tmLocate1) in between calls to the generation
functions.
The lemmas themselves are responsible of registering the names to be
put in the environment.