big update
Implement generation of most of the lemmas in MetaCoq. Works reasonable
well with my custom AST. Code for scoped syntax/variadic syntax/functor
is not generated yet, though. But it's straightforward to extend to
support all that like in the OCaml implementation.

Also rename a lot of the files to make more sense.