better handling of implicit arguments
Before we always assumed implicit arguments are at the front of the
argument list. However this assumption is wrong with several lemmas.
We could reorder the argument lists but that would only delay problems.
So now for each constant we generate we save a bitfield indicating which
arguments are implicit. When we translate an application of this
constant to arguments, we intersperse the given arguments with holes in
the appropriate places.